Source refactor
For easier maintenance, this checkin just moves things around;
Ssplits into multiple source files, and adds 'setup.py' to make
an installable package with a command-line entrypoint 'z80conv'.
Also adds setup.cfg settings for pycodestyle and pylint,
and an explicit license file with the GPLv2 text.

For development, it will be convenient to install using
  pip install -e <directory>
then you can run the current code using `z80conv < FILE.ASM`.
