+++ .gitignore
@@ -305,6 +305,7 @@ paket-files/
 # Python Tools for Visual Studio (PTVS)
 __pycache__/
 *.pyc
+*.egg-info
 
 # Cake - Uncomment if you are using it
 # tools/**
