+++ conv/setup.cfg
@@ -1,0 +1,9 @@
+[pylint.BASIC]
+good-names=op,z1,z2,c,n,e,i,db
+
+[pylint.FORMAT]
+max-line-length=120
+
+[pycodestyle]
+max-line-length=120
+exclude=build,.eggs,.tox,.git,__pycache__,setup.cfg,LICENSE
